Create Control Object Dictionary
